requirements.txt
numpy
scikit-learn
matplotlib
pillow
joblib
torch
torchvision
tqdm
pandas
seaborn
